feat: more argument checks feat: more convenience macros fix: wrong print CHOICE and UINPUT calls fix: missing file unlocks and locks when printing fix: random edge cases change: overall codebase consistency changes change: use bit exact types for almost everything change: split arch into os, memory, compiler, types remove: custom sequence types remove: struct and enum typedefs

+++ code/include/core/io/format/fasta.h
@@ -5,14 +5,16 @@
 #include <stdbool.h>
 #include <stddef.h>
 
+#include "system/types.h"
+
 bool fasta_entry_next(char *restrict *restrict p_cursor);
 
-size_t fasta_total_entries(char *restrict file_cursor, char *restrict file_end);
+u64 fasta_total_entries(char *restrict file_cursor, char *restrict file_end);
 
-size_t fasta_entry_extract(char *restrict *restrict p_cursor,
-			   char *restrict file_end, char *restrict output);
+u64 fasta_entry_extract(char *restrict *restrict p_cursor,
+			char *restrict file_end, char *restrict output);
 
-size_t fasta_entry_length(char *cursor, char *file_end);
+u64 fasta_entry_length(char *cursor, char *file_end);
 
 bool fasta_validate(char *restrict file_start, char *restrict file_end);
