+++ code/include/util/progress.h
@@ -4,10 +4,10 @@
 
 #include <stdatomic.h>
 #include <stdbool.h>
-#include <stddef.h>
 
-bool progress_start(_Atomic(size_t) *progress, size_t total,
-		    const char *message);
+#include "system/types.h"
+
+bool progress_start(_Atomic(u64) *progress, u64 total, const char *message);
 
 void progress_end(void);
